split feature extraction out into a separate file
